# backend/server/api/rest/openapi/dynamic-openapi.yaml
openapi: 3.0.3
info:
  title: BuildBeaver Dynamic Build API - OpenAPI 3.0
  description: |-
    This is the BuildBeaver Dynamic Build API.
  license:
    name: MIT
    url: https://opensource.org/license/mit/
  version: 0.3.00
...
servers:
  - url: http://localhost:3003/api/v1/dynamic
...
paths:
...
  /builds/{buildId}/events:
    get:
      tags:
        - events
      summary: Reads events relating to a build.
      description: Reads events relating to a build that are newer than the last event previously read. Returns immediately if no new events are available.
      operationId: getEvents
      parameters:
        - name: buildId
          in: path
          required: true
          description: The ID of the build to read events for
          schema:
            type: string
          example: 'build:4738115e-070a-44fe-bce0-b43582583eaa'
        - name: last
          in: query
          required: false
          description: The sequence number of the last event previously received; only events with higher sequence numbers than this will be returned.
          schema:
            type: integer
            format: int64
          example: 350
        - name: limit
          in: query
          required: false
          description: The maximum number of events to return
          schema:
            type: integer
            format: int64
          example: '100'
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Events'
        '400':
          description: Invalid input
      security:
        - jwt_build_token: []
...
  /artifacts/{artifactId}/data:
    get:
      tags:
        - build
      summary: Reads the data for an artifact.
      description: Reads the data for a particular artifact.
      operationId: getArtifactData
      parameters:
        - name: artifactId
          in: path
          required: true
          description: The ID of the artifact to read information about.
          schema:
            type: string
          example: 'artifact:5238115e-070a-44fe-bce0-b43582583eff'
      responses:
        '200':
          description: Successfully fetched the data for an artifact
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
      security:
        - jwt_build_token: []
...
components:
  schemas:
...
    Events:
      type: array
      items:
        $ref: '#/components/schemas/Event'
    Event:
      type: object
      required:
        - id
        - created_at
        - sequence_number
        - build_id
        - type
        - resource_id
        - resource_name
        - payload
      properties:
        # Metadata
        id:
          type: string
        created_at:
          type: string
          format: date-time
        # Data
        sequence_number:
          type: integer
          format: int64
          description: SequenceNumber is a monotonically increasing number to provide a well-defined order for events within a build. Event sequence numbers will normally (but not always) be contiguous for a given build.
        build_id:
          type: string
          description: The ID of the build to fetch events for.
          example: 'build:4738115e-070a-44fe-bce0-b43582583eaa'
        type:
          type: string
          description: The type of event, determining what is expected in the event data.
          enum:
            - BuildStatusChanged
            - JobStatusChanged
            - StepStatusChanged
        resource_id:
          type: string
          description: The ID of the resource this event is associated with.
          example: 'job:2238115e-070a-44fe-bce0-b43582583edd'
        workflow:
          type: string
          description: The name of the workflow this event is associated with, if applicable
          example: 'front-end-workflow'
        job_name:
          type: string
          description: The name of the job this event relates to, if applicable
          example: 'generate-code'
        resource_name:
          type: string
          description: The name of the resource this event is associated with. For job-related events this will be the same as job_name.
          example: 'generate-code'
        payload:
          type: string
          description: Payload provides additional information for the event. The format of the payload data depends on the event Type.

    # Log-related schemas
...
  securitySchemes:
    jwt_build_token:
      type: apiKey
      name: Authorization
      in: header